chore: WebApplicationManager code quality changes
- Created HttpResultTools to reduce core WebApplicationManager
- Created CallbackCallerTools to reduce core WebApplicationManager
